pyproject.toml
[project]
name = "sir-convert-a-lot"
version = "0.1.0"
description = "Standalone, LLM-friendly conversion platform with canonical CLI/API surfaces."
authors = [
    { name = "Olof Larsson", email = "paunchygent@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.128.8",
    "uvicorn>=0.40.0",
    "python-multipart>=0.0.22",
    "prometheus-client>=0.23.1",
    "httpx>=0.28.1",
    "typer>=0.16.0",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.3",
    "mdformat>=1.0.0",
    "mdformat-gfm>=0.4.1",
    "weasyprint>=63.0",
    "docling==2.73.1",
    "pymupdf4llm==0.3.4",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.14.0",
    "mypy>=1.18.2",
    "types-pyyaml>=6.0.12.20250915",
    "mdformat-frontmatter>=2.0.10",
]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
]

[tool.ruff.lint]
select = ["E", "W", "F", "I"]
ignore = ["E203"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_return_any = true
check_untyped_defs = true
ignore_missing_imports = true
explicit_package_bases = true
files = [
    "scripts/sir_convert_a_lot",
    "scripts/docs_as_code",
    "tests/sir_convert_a_lot",
]

[tool.pytest.ini_options]
addopts = "-s -v"
python_files = "test_*.py *_test.py"
pythonpath = "."
testpaths = ["tests"]
filterwarnings = [
    "ignore:builtin type SwigPyPacked has no __module__ attribute:DeprecationWarning",
    "ignore:builtin type SwigPyObject has no __module__ attribute:DeprecationWarning",
    "ignore:builtin type swigvarlink has no __module__ attribute:DeprecationWarning",
]

[tool.pdm]
distribution = false

[tool.pdm.scripts]
"serve:sir-convert-a-lot" = "uvicorn scripts.sir_convert_a_lot.service:app --host 0.0.0.0 --port 8085"
"serve:sir-convert-a-lot-eval" = "uvicorn scripts.sir_convert_a_lot.service_eval:app --host 0.0.0.0 --port 8086"
"convert-a-lot" = "python -m scripts.sir_convert_a_lot.cli"
"sir-convert-a-lot" = "python -m scripts.sir_convert_a_lot.cli"
"benchmark:story-003b" = "python -m scripts.sir_convert_a_lot.benchmark_gpu_governance"
"benchmark:task-12" = "python -m scripts.sir_convert_a_lot.benchmark_scientific_corpus"
"validate:docling-gpu-live" = "python -m scripts.sir_convert_a_lot.live_docling_gpu_quality"
"run-local-pdm" = "bash scripts/devops/run-local-pdm.sh"
"run-hemma" = "bash scripts/devops/run-hemma.sh"
"hemma-verify-gpu-runtime" = "bash scripts/devops/verify-hemma-gpu-runtime.sh"
"hemma-verify-v2-conversions" = "bash scripts/devops/verify-hemma-v2-conversions.sh"
"hemma-repair-rocm-runtime" = "bash scripts/devops/repair-hemma-rocm-runtime.sh"
"dev-start" = "bash scripts/devops/dev-compose.sh start"
"dev-stop" = "bash scripts/devops/dev-compose.sh stop"
"dev-build" = "bash scripts/devops/dev-compose.sh build"
"dev-build-clean" = "bash scripts/devops/dev-compose.sh build-clean"
"dev-recreate" = "bash scripts/devops/dev-compose.sh recreate"
"dev-logs" = "bash scripts/devops/dev-compose.sh logs"
"dev-ps" = "bash scripts/devops/dev-compose.sh ps"
"dev-config" = "bash scripts/devops/dev-compose.sh config"
"dev-check" = "bash scripts/devops/dev-compose.sh check"
"new-task" = "python -m scripts.docs_as_code.new_task"
"new-programme" = "python -m scripts.docs_as_code.new_task --type programme"
"new-epic" = "python -m scripts.docs_as_code.new_task --type epic"
"new-story" = "python -m scripts.docs_as_code.new_task --type story"
"new-review" = "python -m scripts.docs_as_code.new_task --type review"
"new-fix" = "python -m scripts.docs_as_code.new_task --type fix"
"new-doc" = "python -m scripts.docs_as_code.new_doc"
"new-rule" = "python -m scripts.docs_as_code.new_rule"
"validate-tasks" = "python -m scripts.docs_as_code.validate_tasks"
"validate-docs" = "python -m scripts.docs_as_code.validate_docs"
"docs-validate" = "python -m scripts.docs_as_code.validate_docs"
"index-tasks" = "python -m scripts.docs_as_code.index_tasks"
"format" = "ruff format ."
"format-md" = "mdformat --extensions frontmatter README.md AGENTS.md docs .agents"
"format-all" = { composite = [
    "ruff format .",
    "mdformat --extensions frontmatter README.md AGENTS.md docs .agents",
] }
"lint" = { composite = [
    "ruff format --check --diff .",
    "ruff check .",
    "mdformat --check --extensions frontmatter README.md AGENTS.md docs .agents",
    "python -m scripts.docs_as_code.validate_docs",
    "python -m scripts.docs_as_code.validate_tasks",
] }
"lint-fix" = { composite = [
    "ruff check --fix --unsafe-fixes .",
    "ruff format .",
    "mdformat --extensions frontmatter README.md AGENTS.md docs .agents",
    "python -m scripts.docs_as_code.validate_docs",
    "python -m scripts.docs_as_code.validate_tasks",
] }
"typecheck" = "mypy --config-file pyproject.toml"
"typecheck-all" = "mypy --config-file pyproject.toml"
"test" = "pytest -s"
"pytest-root" = "pytest -s"

[tool.sir_convert_a_lot.rocm_runtime]
torch_index_url = "https://download.pytorch.org/whl/rocm7.1"
torch_version = "2.10.0+rocm7.1"
torchvision_version = "0.25.0+rocm7.1"
torchaudio_version = "2.10.0+rocm7.1"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
